Add ScrollBtn Module/Fix Linters
